Reorg in preparation for UniFFI client.
