+++ cspell.json
@@ -14,6 +14,8 @@
         "nanocore",
         "numpy",
         "graphviz",
+        "uniffi",
+        "cffi"
     ],
     "ignoreWords": [
         "relpath",
