+++ src/core/crypto.rs
@@ -1,7 +1,9 @@
 use crate::{
-    error::Result,
-    model::{Blob, BlobKind},
-    util::get,
+    core::util::get,
+    uniffi::{
+        error::Result,
+        model::{Blob, BlobKind},
+    },
 };
 use serde_yaml;
 use sha2::{Digest as _, Sha256};
@@ -21,7 +23,7 @@ use std::{
     clippy::indexing_slicing,
     reason = "Reading less than 0 is impossible."
 )]
-pub fn hash_stream(stream: &mut impl Read) -> Result<String> {
+pub(crate) fn hash_stream(stream: &mut impl Read) -> Result<String> {
     const BUFFER_SIZE: usize = 8 << 10; // 8KB chunks to match with page size typically found
     let mut hash = Sha256::new();
 
@@ -79,7 +81,7 @@ pub fn hash_dir(dirpath: impl AsRef<Path>) -> Result<String> {
 /// # Errors
 ///
 /// Will return error if hashing fails on file or directory.
-pub fn hash_blob(
+pub(crate) fn hash_blob(
     namespace_lookup: &HashMap<String, PathBuf, RandomState>,
     blob: Blob,
 ) -> Result<Blob> {
