+++ src/core/mod.rs
@@ -1,0 +1,8 @@
+/// State change verification via cryptographic utilities.
+pub mod crypto;
+pub(crate) mod error;
+/// Components of the data model.
+pub mod model;
+pub(crate) mod orchestrator;
+pub(crate) mod store;
+pub(crate) mod util;
