+++ src/core/store/filestore.rs
@@ -1,8 +1,10 @@
 use crate::{
-    error::{Kind, OrcaError, Result},
-    model::{Annotation, Pod, PodJob, PodResult, to_yaml},
-    store::{ModelID, ModelInfo, Store},
-    util::get_type_name,
+    core::{error::Kind, model::to_yaml, util::get_type_name},
+    uniffi::{
+        error::{OrcaError, Result},
+        model::Annotation,
+        store::{ModelID, ModelInfo, Store as _, filestore::LocalFileStore},
+    },
 };
 use colored::Colorize as _;
 use glob::glob;
@@ -15,72 +17,6 @@ use std::{
     path::{Path, PathBuf},
     sync::LazyLock,
 };
-/// Support for a storage backend on a local filesystem directory.
-#[derive(Debug)]
-pub struct LocalFileStore {
-    /// A local path to a directory where store will be located.
-    directory: PathBuf,
-}
-
-impl Store for LocalFileStore {
-    fn save_pod(&self, pod: &Pod) -> Result<()> {
-        self.save_model(pod, &pod.hash, pod.annotation.as_ref())
-    }
-    fn load_pod(&self, model_id: &ModelID) -> Result<Pod> {
-        let (mut pod, annotation, hash) = self.load_model::<Pod>(model_id)?;
-        pod.annotation = annotation;
-        pod.hash = hash;
-        Ok(pod)
-    }
-    fn list_pod(&self) -> Result<Vec<ModelInfo>> {
-        self.list_model::<Pod>()
-    }
-    fn delete_pod(&self, model_id: &ModelID) -> Result<()> {
-        self.delete_model::<Pod>(model_id)
-    }
-    fn save_pod_job(&self, pod_job: &PodJob) -> Result<()> {
-        self.save_pod(&pod_job.pod)?;
-        self.save_model(pod_job, &pod_job.hash, pod_job.annotation.as_ref())
-    }
-    fn load_pod_job(&self, model_id: &ModelID) -> Result<PodJob> {
-        let (mut pod_job, annotation, hash) = self.load_model::<PodJob>(model_id)?;
-        pod_job.annotation = annotation;
-        pod_job.hash = hash;
-        pod_job.pod = self.load_pod(&ModelID::Hash(pod_job.pod.hash))?;
-        Ok(pod_job)
-    }
-    fn list_pod_job(&self) -> Result<Vec<ModelInfo>> {
-        self.list_model::<PodJob>()
-    }
-    fn delete_pod_job(&self, model_id: &ModelID) -> Result<()> {
-        self.delete_model::<PodJob>(model_id)
-    }
-    fn save_pod_result(&self, pod_result: &PodResult) -> Result<()> {
-        self.save_pod_job(&pod_result.pod_job)?;
-        self.save_model(pod_result, &pod_result.hash, pod_result.annotation.as_ref())
-    }
-    fn load_pod_result(&self, model_id: &ModelID) -> Result<PodResult> {
-        let (mut pod_result, annotation, hash) = self.load_model::<PodResult>(model_id)?;
-        pod_result.annotation = annotation;
-        pod_result.hash = hash;
-        pod_result.pod_job = self.load_pod_job(&ModelID::Hash(pod_result.pod_job.hash))?;
-        Ok(pod_result)
-    }
-    fn list_pod_result(&self) -> Result<Vec<ModelInfo>> {
-        self.list_model::<PodResult>()
-    }
-    fn delete_pod_result(&self, model_id: &ModelID) -> Result<()> {
-        self.delete_model::<PodResult>(model_id)
-    }
-    fn delete_annotation<T>(&self, name: &str, version: &str) -> Result<()> {
-        let hash = self.lookup_hash::<T>(name, version)?;
-        let annotation_file =
-            self.make_path::<T>(&hash, Self::make_annotation_relpath(name, version));
-        fs::remove_file(&annotation_file)?;
-
-        Ok(())
-    }
-}
 
 #[expect(clippy::expect_used, reason = "Valid static regex")]
 static RE_MODEL_METADATA: LazyLock<Regex> = LazyLock::new(|| {
@@ -117,7 +53,7 @@ impl LocalFileStore {
     pub fn make_path<T>(&self, hash: &str, relpath: impl AsRef<Path>) -> PathBuf {
         PathBuf::from(format!(
             "{}/{}/{}/{}",
-            self.directory.to_string_lossy(),
+            self.get_directory().to_string_lossy(),
             Self::MODEL_NAMESPACE,
             get_type_name::<T>().to_snake_case(),
             hash
@@ -139,18 +75,12 @@ impl LocalFileStore {
         });
         Ok(paths)
     }
-    /// Get the directory where store is located.
-    pub fn get_directory(&self) -> &Path {
-        &self.directory
-    }
-    /// Construct a local file store instance in a specific directory.
-    pub fn new(directory: impl AsRef<Path>) -> Self {
-        Self {
-            directory: directory.as_ref().into(),
-        }
-    }
-
-    fn lookup_hash<T>(&self, name: &str, version: &str) -> Result<String> {
+    /// Find hash using name and version.
+    ///
+    /// # Errors
+    ///
+    /// Will return error if unable to find.
+    pub(crate) fn lookup_hash<T>(&self, name: &str, version: &str) -> Result<String> {
         let model_info = Self::find_model_metadata(
             &self.make_path::<T>("*", Self::make_annotation_relpath(name, version)),
         )?
@@ -172,8 +102,12 @@ impl LocalFileStore {
         fs::write(file, content)?;
         Ok(())
     }
-
-    fn save_model<T: Serialize>(
+    /// How any model is stored.
+    ///
+    /// # Errors
+    ///
+    /// Will return `Err` if there is an issue storing the model.
+    pub(crate) fn save_model<T: Serialize>(
         &self,
         model: &T,
         hash: &str,
@@ -220,8 +154,13 @@ impl LocalFileStore {
         }
         Ok(())
     }
-
-    fn load_model<T: DeserializeOwned>(
+    /// How to load any stored model into an instance.
+    ///
+    /// # Errors
+    ///
+    /// Will return `Err` if there is an issue loading the model from the store using `name` and
+    /// `version`.
+    pub(crate) fn load_model<T: DeserializeOwned>(
         &self,
         model_id: &ModelID,
     ) -> Result<(T, Option<Annotation>, String)> {
@@ -247,12 +186,21 @@ impl LocalFileStore {
             }
         }
     }
-
-    fn list_model<T>(&self) -> Result<Vec<ModelInfo>> {
+    /// How to query any stored models.
+    ///
+    /// # Errors
+    ///
+    /// Will return `Err` if there is an issue querying metadata from existing models in the store.
+    pub(crate) fn list_model<T>(&self) -> Result<Vec<ModelInfo>> {
         Ok(Self::find_model_metadata(&self.make_path::<T>("**", "*"))?.collect())
     }
-
-    fn delete_model<T>(&self, model_id: &ModelID) -> Result<()> {
+    /// How to explicitly delete any stored model and all associated annotations (does not propagate).
+    ///
+    /// # Errors
+    ///
+    /// Will return `Err` if there is an issue deleting a model from the store using `name` and
+    /// `version`.
+    pub(crate) fn delete_model<T>(&self, model_id: &ModelID) -> Result<()> {
         // assumes propagate = false
         let hash = match model_id {
             ModelID::Hash(hash) => hash,
