+++ src/uniffi/mod.rs
@@ -1,0 +1,8 @@
+/// Error handling.
+pub mod error;
+/// Components of the data model.
+pub mod model;
+/// Interface into container orchestration engine.
+pub mod orchestrator;
+/// Data persistence provided by a store backend.
+pub mod store;
