+++ src/uniffi/orchestrator/mod.rs
@@ -1,7 +1,6 @@
-use crate::{
+use crate::uniffi::{
     error::Result,
     model::{OrcaPath, PodJob, PodResult},
-    util::get_type_name,
 };
 use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, future::Future, path::PathBuf};
@@ -56,16 +55,6 @@ pub struct PodRun {
     pub orchestrator_source: String,
     /// Name given by orchestrator.
     pub assigned_name: String,
-}
-
-impl PodRun {
-    fn new<O: Orchestrator>(pod_job: &PodJob, assigned_name: String) -> Self {
-        Self {
-            pod_job: pod_job.clone(),
-            orchestrator_source: get_type_name::<O>(),
-            assigned_name,
-        }
-    }
 }
 
 /// API for standard behavior of any container orchestration engine supported.
