+++ tests/model.rs
@@ -3,7 +3,7 @@
 pub mod fixture;
 use fixture::{NAMESPACE_LOOKUP_READ_ONLY, pod_job_style, pod_result_style, pod_style};
 use indoc::indoc;
-use orcapod::{error::Result, model::to_yaml};
+use orcapod::{core::model::to_yaml, uniffi::error::Result};
 
 #[test]
 fn hash_pod() -> Result<()> {
